requirements.txt
cryptg
FastTelethonhelper